+++ common/logger.go
@@ -98,13 +98,3 @@ func WaitLogs() {
 	runtime.Gosched()
 	logbarrier.Wait()
 }
-
-func Progress(step int, form string, details ...interface{}) {
-	previous := ProgressMark
-	ProgressMark = time.Now()
-	delta := ProgressMark.Sub(previous).Round(1 * time.Millisecond).Seconds()
-	message := fmt.Sprintf(form, details...)
-	Log("####  Progress: %02d/15  %s  %8.3fs  %s", step, Version, delta, message)
-	Timeline("%d/15 %s", step, message)
-	RunJournal("environment", "build", "Progress: %02d/15  %s  %8.3fs  %s", step, Version, delta, message)
-}